requirements.txt
flask
opencv-python
mediapipe
matplotlib
numpy
Pillow
requests